# Conflicts:
#	contracts/azorius/LinearERC20Voting.sol
#	test/Azorius-LinearERC20Voting.test.ts
#	test/AzoriusFreezeGuard-ERC20FreezeVoting.test.ts
#	test/AzoriusFreezeGuard-MultisigFreezeVoting.test.ts

+++ contracts/ERC20FreezeVoting.sol
@@ -1,9 +1,9 @@
 //SPDX-License-Identifier: MIT
 pragma solidity =0.8.19;
 
-import "./BaseFreezeVoting.sol";
-import "@gnosis.pm/safe-contracts/contracts/common/Enum.sol";
-import "@openzeppelin/contracts/governance/utils/IVotes.sol";
+import { BaseFreezeVoting, IBaseFreezeVoting } from "./BaseFreezeVoting.sol";
+import { Enum } from "@gnosis.pm/safe-contracts/contracts/common/Enum.sol";
+import { IVotes } from "@openzeppelin/contracts/governance/utils/IVotes.sol";
 
 /**
  * A [BaseFreezeVoting](./BaseFreezeVoting.md) implementation which handles 
@@ -33,12 +33,12 @@ contract ERC20FreezeVoting is BaseFreezeVoting {
         (
             address _owner,
             uint256 _freezeVotesThreshold,
-            uint256 _freezeProposalPeriod,
-            uint256 _freezePeriod,
+            uint32 _freezeProposalPeriod,
+            uint32 _freezePeriod,
             address _votesERC20
         ) = abi.decode(
                 initializeParams,
-                (address, uint256, uint256, uint256, address)
+                (address, uint256, uint32, uint32, address)
             );
 
         __Ownable_init();
@@ -59,7 +59,7 @@ contract ERC20FreezeVoting is BaseFreezeVoting {
         if (block.number > freezeProposalCreatedBlock + freezeProposalPeriod) {
             // create a new freeze proposal and set total votes to msg.sender's vote count
 
-            freezeProposalCreatedBlock = block.number;
+            freezeProposalCreatedBlock = uint32(block.number);
 
             userVotes = votesERC20.getPastVotes(
                 msg.sender,
