+++ contracts/interfaces/IMultisigFreezeGuard.sol
@@ -1,7 +1,7 @@
 //SPDX-License-Identifier: MIT
 pragma solidity =0.8.19;
 
-import "@gnosis.pm/safe-contracts/contracts/common/Enum.sol";
+import { Enum } from "@gnosis.pm/safe-contracts/contracts/common/Enum.sol";
 
 /**
  * A specification for a Safe Guard contract which allows for multi-sig DAOs (Safes)
@@ -61,7 +61,7 @@ interface IMultisigFreezeGuard {
      *
      * @param _timelockPeriod new timelock period for the subDAO (in blocks)
      */
-    function updateTimelockPeriod(uint256 _timelockPeriod) external;
+    function updateTimelockPeriod(uint32 _timelockPeriod) external;
 
     /**
      * Updates the execution period.
@@ -73,13 +73,13 @@ interface IMultisigFreezeGuard {
      *
      * @param _executionPeriod number of blocks a transaction has to be executed within
      */
-    function updateExecutionPeriod(uint256 _executionPeriod) external;
+    function updateExecutionPeriod(uint32 _executionPeriod) external;
 
     /**
      * Gets the block number that the given transaction was timelocked at.
      *
      * @param _transactionHash hash of the transaction data
-     * @return uint256 block number in which the transaction began its timelock period
+     * @return uint32 block number in which the transaction began its timelock period
      */
-    function getTransactionTimelockedBlock(bytes32 _transactionHash) external view returns (uint256);
+    function getTransactionTimelockedBlock(bytes32 _transactionHash) external view returns (uint32);
 }
