+++ test/MultisigFreezeGuard-ERC20FreezeVoting.test.ts
@@ -137,7 +137,7 @@ describe("Child Multisig DAO with Azorius Parent", () => {
 
     // Deploy MultisigFreezeGuard contract with a 60 block timelock period, and a 60 block execution period
     const freezeGuardSetupData = abiCoder.encode(
-      ["uint256", "uint256", "address", "address", "address"],
+      ["uint32", "uint32", "address", "address", "address"],
       [
         60, // Timelock period
         60, // Execution period
@@ -151,7 +151,7 @@ describe("Child Multisig DAO with Azorius Parent", () => {
 
     // Initialize FreezeVoting contract
     const freezeVotingSetupData = abiCoder.encode(
-      ["address", "uint256", "uint256", "uint256", "address", "address"],
+      ["address", "uint256", "uint32", "uint32", "address", "address"],
       [
         freezeGuardOwner.address,
         1090, // freeze votes threshold
